+++ src/templates/bespoke/utils.ts
@@ -5,6 +5,13 @@ type LocationLike = Pick<
 type QuerySetter = (...args: Parameters<History['pushState']>) => void
 
 const replacer: QuerySetter = (...args) => history.replaceState(...args)
+const viewAttr = 'data-bespoke-view'
+
+export enum ViewMode {
+  Normal = '',
+  Presenter = 'presenter',
+  Next = 'next',
+}
 
 export const generateURLfromParams = (
   params: URLSearchParams,
@@ -12,6 +19,19 @@ export const generateURLfromParams = (
 ) => {
   const q = params.toString()
   return `${protocol}//${host}${pathname}${q ? '?' : ''}${q}${hash}`
+}
+
+export const getViewMode = () => {
+  switch (document.body.getAttribute(viewAttr)) {
+    case ViewMode.Normal:
+      return ViewMode.Normal
+    case ViewMode.Presenter:
+      return ViewMode.Presenter
+    case ViewMode.Next:
+      return ViewMode.Next
+    default:
+      throw new Error('View mode is not assigned.')
+  }
 }
 
 export const readQuery = (name: string) =>
@@ -52,3 +72,18 @@ export const setQuery = (
     console.error(e)
   }
 }
+
+export const setViewMode = () =>
+  document.body.setAttribute(
+    viewAttr,
+    ((): ViewMode => {
+      switch (readQuery('view')) {
+        case 'next':
+          return ViewMode.Next
+        case 'presenter':
+          return ViewMode.Presenter
+        default:
+          return ViewMode.Normal
+      }
+    })()
+  )